+++ src/test/java/io/github/hectorvent/floci/config/TlsCertificateHostnameTest.java
@@ -476,90 +476,6 @@ class TlsCertificateHostnameTest {
         leaf.verify(ca.getPublicKey());
     }
 
-    // ==================== Runtime-Learned Hostname Tests ====================
-
-    @Test
-    void learnedHostnamesAreKeptWhenBootRegeneratesTheCertificate() throws Exception {
-        System.setProperty("floci.hostname", "host1");
-        System.setProperty("floci.tls.enabled", "true");
-        System.setProperty("floci.tls.self-signed", "true");
-        System.setProperty("floci.storage.persistent-path", tempDir.toString());
-        new TlsConfigSource();
-
-        Path metadataFile = tempDir.resolve("tls/floci-server.metadata.json");
-        var mapper = new com.fasterxml.jackson.databind.ObjectMapper();
-        CertificateMetadata metadata = mapper.readValue(metadataFile.toFile(), CertificateMetadata.class);
-        metadata.setLearnedHostnames(List.of("api.example.localhost.floci.io"));
-        Files.writeString(metadataFile, mapper.writeValueAsString(metadata));
-
-        System.setProperty("floci.hostname", "host2");
-        new TlsConfigSource();
-
-        List<String> sans = extractSansFromCertificate(parseCertificate(tempDir.resolve("tls/floci-server.crt")));
-        assertTrue(sans.contains("host2"), sans.toString());
-        assertFalse(sans.contains("host1"), sans.toString());
-        assertTrue(sans.contains("api.example.localhost.floci.io"), "learned name must survive: " + sans);
-        CertificateMetadata after = mapper.readValue(metadataFile.toFile(), CertificateMetadata.class);
-        assertEquals(List.of("api.example.localhost.floci.io"), after.getLearnedHostnames());
-        assertFalse(after.getHostnames().contains("api.example.localhost.floci.io"),
-                "learned names stay out of the configured list");
-    }
-
-    @Test
-    void learnedHostnamesAreKeptWhenAnExpiredLeafIsReissued() throws Exception {
-        System.setProperty("floci.tls.enabled", "true");
-        System.setProperty("floci.tls.self-signed", "true");
-        System.setProperty("floci.storage.persistent-path", tempDir.toString());
-        Path tlsDir = Files.createDirectories(tempDir.resolve("tls"));
-        FlociCertificateAuthority ca = FlociCertificateAuthority.loadOrCreate(tlsDir);
-        CertificateGenerator gen = new CertificateGenerator();
-        List<String> configured = List.of("localhost", "127.0.0.1", "0.0.0.0", "*.localhost",
-                "localhost.floci.io", "*.localhost.floci.io", "*.execute-api.localhost.floci.io",
-                "*.execute-api.localhost.localstack.cloud", "host.docker.internal");
-        List<String> sans = new ArrayList<>(configured);
-        sans.add("api.example.localhost.floci.io");
-        java.security.KeyPair keyPair = java.security.KeyPairGenerator.getInstance("RSA").generateKeyPair();
-        X509Certificate expired = gen.signCertificate(new org.bouncycastle.asn1.x500.X500Name("CN=localhost"),
-                keyPair.getPublic(), org.bouncycastle.asn1.x500.X500Name.getInstance(
-                        ca.certificate().getSubjectX500Principal().getEncoded()), ca.key(), sans, false,
-                CertificateGenerator.LeafUsage.SERVER, -1);
-        Files.writeString(tlsDir.resolve("floci-server.crt"), gen.toPem(expired));
-        Files.writeString(tlsDir.resolve("floci-server.key"), gen.toPem(keyPair.getPrivate()));
-        CertificateMetadata metadata = CertificateMetadata.create(configured, "dev");
-        metadata.setLearnedHostnames(List.of("api.example.localhost.floci.io"));
-        Files.writeString(tlsDir.resolve("floci-server.metadata.json"),
-                new com.fasterxml.jackson.databind.ObjectMapper().writeValueAsString(metadata));
-
-        new TlsConfigSource();
-
-        X509Certificate leaf = parseCertificate(tlsDir.resolve("floci-server.crt"));
-        leaf.checkValidity();
-        assertNotEquals(expired.getSerialNumber(), leaf.getSerialNumber(), "the expired leaf was reissued");
-        List<String> reissued = extractSansFromCertificate(leaf);
-        assertTrue(reissued.contains("api.example.localhost.floci.io"), "learned name must survive: " + reissued);
-        assertTrue(reissued.containsAll(configured), reissued.toString());
-    }
-
-    @Test
-    void aLearnedHostnameDoesNotCountAsAConfigurationChange() throws Exception {
-        System.setProperty("floci.tls.enabled", "true");
-        System.setProperty("floci.tls.self-signed", "true");
-        System.setProperty("floci.storage.persistent-path", tempDir.toString());
-        new TlsConfigSource();
-
-        Path metadataFile = tempDir.resolve("tls/floci-server.metadata.json");
-        var mapper = new com.fasterxml.jackson.databind.ObjectMapper();
-        CertificateMetadata metadata = mapper.readValue(metadataFile.toFile(), CertificateMetadata.class);
-        metadata.setLearnedHostnames(List.of("api.example.localhost.floci.io"));
-        Files.writeString(metadataFile, mapper.writeValueAsString(metadata));
-        String certBefore = Files.readString(tempDir.resolve("tls/floci-server.crt"));
-
-        new TlsConfigSource();
-
-        assertEquals(certBefore, Files.readString(tempDir.resolve("tls/floci-server.crt")),
-                "same configuration: the certificate is reused, not regenerated");
-    }
-
     // ==================== Helper Methods ====================
 
     private X509Certificate parseCertificate(Path certFile) throws Exception {
